036 difference-of-squares casting

+++ exercism/projects/java/difference-of-squares/src/main/java/DifferenceOfSquaresCalculator.java
@@ -1,7 +1,7 @@
 class DifferenceOfSquaresCalculator {
 
     int computeSquareOfSumTo(int input) {
-        return (int) Math.pow( input*(input+1)/2, 2);
+        return (int) Math.pow( (double) (input * (input + 1)) /2, 2);
     }
 
     int computeSumOfSquaresTo(int input) {
